+++ test/org/clojars/aldebogdanov/quint_connect/fixtures/args_bad_entry.clj
@@ -1,0 +1,7 @@
+(ns org.clojars.aldebogdanov.quint-connect.fixtures.args-bad-entry
+  "A map entry whose value does not name a pick.")
+
+(defn transfer
+  {:quint/action "transfer" :quint/args [{:from "src"}]}
+  [{:keys [from]}]
+  from)